test: relocate schema suites, seam CLI scripts, and share src builders
H3: redemption-backstop schema/settlement cases and the live-reserve
catalog-integrity cases move into their own owners (live-reserve schema
file 4.00s -> 0.31s). H21: pharos-change-contract readHookInput() and the
CDP mechanism-metrics replay get import-safe seams so tests run in-process
while one real CLI contract remains. H7 (src): status active-price
coverage, dispatch-metadata and mini-app state builders live in their
existing shared fixture owners; every migrated consumer was mutation-checked.
